+++ src/drv_hmc5883l.h
@@ -2,4 +2,4 @@
 
 bool hmc5883lDetect(sensor_align_e align);
 void hmc5883lInit(void);
-void hmc5883lRead(int16_t *magData);
+void hmc5883lRead(int32_t *magData);
