+++ src/utils.h
@@ -22,6 +22,6 @@ void performRotation( rotation_context_st const * const pctx, int32_t *vec );
 
 int constrain(int amt, int low, int high);
 // sensor orientation
-void alignSensors(int16_t *src, int16_t *dest, uint8_t rotation);
+void alignSensors(int32_t *src, int32_t *dest, uint8_t rotation);
 void initBoardAlignment(void);
 void productionDebug(void);
